refactor: 스터디 관련 레포지토리 count 메소드가 long 타입을 반환하도록 변경

+++ src/main/java/com/jj/swm/domain/study/core/dto/response/GetRecruitmentPositionDetailsResponse.java
@@ -8,14 +8,14 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder
 public class GetRecruitmentPositionDetailsResponse extends GetRecruitmentPositionResponse {
 
-    private Integer acceptedCount;
+    private long acceptedCount;
 
-    private Integer participatedCount;
+    private long participatedCount;
 
     public static GetRecruitmentPositionDetailsResponse of(
             StudyRecruitmentPosition recruitmentPosition,
-            Integer acceptedCount,
-            Integer participatedCount
+            long acceptedCount,
+            long participatedCount
     ) {
         return GetRecruitmentPositionDetailsResponse.builder()
                 .recruitmentPositionId(recruitmentPosition.getId())
