+++ src/main/java/com/jj/swm/domain/study/participation/dto/AcceptedStudyParticipationCountInfo.java
@@ -4,5 +4,5 @@ public interface AcceptedStudyParticipationCountInfo {
 
     Long getRecruitmentPositionId();
 
-    Integer getAcceptedStudyParticipationCount();
+    long getAcceptedStudyParticipationCount();
 }
