+++ src/main/java/com/jj/swm/domain/study/participation/dto/StudyParticipationCountInfo.java
@@ -4,7 +4,7 @@ public interface StudyParticipationCountInfo {
 
     Long getRecruitmentPositionId();
 
-    Integer getTotalCount();
+    long getTotalCount();
 
-    Integer getAcceptedCount();
+    long getAcceptedCount();
 }
